LayoutTests/http/tests/inspector/network/network-timing-expected.txt

CONSOLE MESSAGE: line 22: Done.
Tests network panel timing.

http://127.0.0.1:8000/inspector/network/resources/resource.php?type=js&wait=100&send=200&jsdelay=300&jscontent=resourceLoaded()
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=js&wait=100&send=100&jscontent=resourceLoaded()
